+++ test/dts-test.ts
@@ -1,6 +1,6 @@
 import { expectType } from 'tsd';
 import { UAParser, IResult, IBrowser, ICPU, IEngine, IDevice, IOS } from "../src/main/ua-parser";
-import { isAppleSilicon, isChromiumBased } from "../src/helpers/ua-parser-helpers";
+import { isAppleSilicon, isChromeFamily } from "../src/helpers/ua-parser-helpers";
 
 const uastring = 'Mozilla/5.0 (X11; MyCustomOS; Linux i686; rv:19.0) Gecko/20100101 Firefox/19.0';
 const extensions = {
@@ -28,12 +28,12 @@ expectType<IBrowser>(browser);
 expectType<string | undefined>(browser.name);
 expectType<string | undefined>(browser.version);
 expectType<string | undefined>(browser.major);
-expectType<'crawler' | 'cli' | 'email' | 'fetcher' | 'inapp' | 'module' | undefined>(browser.type);
+expectType<'crawler' | 'cli' | 'email' | 'fetcher' | 'inapp' | 'mediaplayer' | 'library' | undefined>(browser.type);
 expectType<boolean>(browser.is(''));
 expectType<string>(browser.toString());
 expectType<IBrowser | PromiseLike<IBrowser>>(browser.withClientHints());
-expectType<IBrowser>((<IBrowser>browser.withClientHints()).withFeatureCheck());
-expectType<boolean>((<IBrowser>browser.withClientHints()).withFeatureCheck().is(''));
+expectType<IBrowser | PromiseLike<IBrowser>>((<IBrowser>browser.withClientHints()).withFeatureCheck());
+expectType<boolean>((<IBrowser>(<IBrowser>browser.withClientHints()).withFeatureCheck()).is(''));
 
 expectType<ICPU>(parser.getCPU());
 expectType<IDevice>(parser.getDevice());
@@ -46,4 +46,4 @@ expectType<UAParser>(parser.setUA(uastring));
 const result = parser.getResult();
 
 expectType<boolean>(isAppleSilicon(result));
-expectType<boolean>(isChromiumBased(result));
+expectType<boolean>(isChromeFamily(result));